+++ .editorconfig
@@ -10,3 +10,6 @@ trim_trailing_whitespace = true
 
 [*.go]
 indent_style = tab
+
+[Makefile]
+indent_style = tab
