+++ CHANGELOG.md
@@ -1,0 +1,1 @@
+# Ranger Changelog
