+++ MAINTAINERS.md
@@ -1,0 +1,8 @@
+# Ranger Maintainers
+
+## Maintainers
+
+| Maintainer | GitHub ID | Affiliation |
+| --------------- | --------- | ----------- |
+| Craig Anderson | [cda0](https://github.com/cda0) | [Frontier Digital](https://github.com/frontierdigital/) |
+| Fraser Davidson | [frasdav](https://github.com/frasdav) | [Frontier Digital](https://github.com/frontierdigital/) |
