+++ Makefile
@@ -1,0 +1,5 @@
+build:
+	go build cmd/ranger/ranger.go
+
+test:
+	go test -v ./...
