+++ README.md
@@ -1,9 +1,1 @@
 # ranger
-
-## Contributing
-
-1. Ensure the correct version of Golang is installed. The version can be found in `./go.mod`.
-1. Install the dependencies with `go install`.
-1. Compile the code with `go build`.
-1. Print the help text with `ranger --help`.
-1. Run the tests with `go test -v ./...`.
