+++ SUPPORT.md
@@ -1,0 +1,3 @@
+# Ranger Support
+
+Thanks for trying out Ranger! We're setting up an IRC channel, but please raise a GitHub issue in the meantime if you've got any questions.
