+++ manifest.yml
@@ -1,0 +1,17 @@
+---
+version: 67
+environment: dev
+layer: infra
+workloads:
+  - name: bootstrap
+    source: Demo/bootstrap-workload
+    version: 1.0.50
+  - name: security-centre
+    source: Demo/security-centre-workload
+    version: 1.0.11
+  - name: virtual-network
+    source: Demo/virtual-network-workload
+    version: 1.0.18
+  - name: rbac-assignments
+    source: Demo/rbac-assignments-workload
+    version: 1.0.7
