+++ pkg/cmd/cli/version/version.go
@@ -1,10 +1,9 @@
 package version
 
 import (
+	"github.com/frontierdigital/ranger/pkg/util/output"
 	"github.com/spf13/cobra"
 	goVersion "go.hein.dev/go-version"
-
-	"github.com/frontierdigital/ranger/core/output"
 )
 
 var (
