+++ pkg/cmd/cli/version/version_test.go
