+++ pkg/util/output/output_test.go
@@ -1,0 +1,1 @@
+package output
